test: move runtime artifact validation coverage
